project/project/requirements.txt
asgiref==3.4.1
better-profanity==0.7.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2.11
django-ckeditor==6.2.0
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
